\name{map}
\title{map}
\alias{map}
\description{Function to draw a map to perform BML model}
\usage{
map(p,pc,r,c)
}
\arguments{
	  \item{p}{(rho) is the percent of available spaces to be covered by cars.}
	  \item{pc}{the percent of the cars that are red, 1-pc percent of the cars will be blue.}
	  \item{r}{the number of columns of the map}
	  \item{c}{the number of rows of the map}
}
